% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde1d.R
\name{bw_kde1d}
\alias{bw_kde1d}
\title{Bandwidth selection for kde1d}
\usage{
bw_kde1d(data, xmin, xmax, n.subs)
}
\arguments{
\item{data}{data vector.}

\item{xmin}{lower bound for the support of the density.}

\item{xmax}{upper bound for the support of the density.}

\item{n.subs}{number of subsamples used to estimate the cross-validation
criterion.}
}
\description{
Bandwidth selection for kde1d
}

