% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde1d.R
\name{kde1d}
\alias{kde1d}
\alias{pkde1d}
\alias{qkde1d}
\alias{rkde1d}
\title{Univariate kernel density estimation for bounded and unbounded support}
\usage{
kde1d(x, mult = 1, xmin = -Inf, xmax = Inf, bw = NULL, bw_min = 0,
  ...)

pkde1d(x, obj)

qkde1d(x, obj)

rkde1d(n, obj, quasi = FALSE)
}
\arguments{
\item{x}{vdeprecated.}

\item{mult}{deprecated.}

\item{xmin}{deprecated.}

\item{xmax}{deprecated..}

\item{bw}{deprecated.}

\item{bw_min}{deprecated.}

\item{...}{deprecated.}

\item{obj}{deprecated.}

\item{n}{deprecated.}

\item{quasi}{deprecated.}
}
\description{
Function is deprecated, use \code{\link[kde1d:kde1d]{kde1d::kde1d()}} instead.
}
