% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde1d.R
\name{plot.kde1d}
\alias{plot.kde1d}
\alias{lines.kde1d}
\alias{lines.kde1d}
\title{Plotting kde1d objects}
\usage{
\method{plot}{kde1d}(x, ...)

\method{lines}{kde1d}(x, ...)
}
\arguments{
\item{x}{\code{kde1d} object.}

\item{...}{further arguments passed to \code{\link{plot.default}}.}
}
\description{
Plotting kde1d objects
}
\examples{
data(wdbc)  # load data
fit <- kde1d(wdbc[, 7])  # estimate density
plot(fit)  # plot density estimate

fit2 <- kde1d(as.ordered(wdbc[, 1])) # discrete variable
plot(fit2, col = 2)

}
\seealso{
\code{\link{kde1d}}
\code{\link{lines.kde1d}}
}
