% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdevine.R
\name{rkdevine}
\alias{rkdevine}
\title{Simulate from a kdevine object}
\usage{
rkdevine(n, obj, quasi = FALSE)
}
\arguments{
\item{n}{number of observations.}

\item{obj}{a \code{kdevine} object.}

\item{quasi}{logical; the default (\code{FALSE}) returns pseudo-random
numbers, use \code{TRUE} for quasi-random numbers (generalized Halton, see
\code{\link[qrng:ghalton]{ghalton}}).}
}
\value{
An \eqn{n x d} matrix of simulated data from the \code{kdevine}
object.
}
\description{
Simulate from a kdevine object
}
\examples{
data(wdbc)  # load data
dat <- wdbc[, 5:7]
pairs(dat)  # plot data

# estimate density (use xmin to indicate positive support)
fit <- kdevine(wdbc[, 5:7], xmin = rep(0, 3))
pairs(rkdevine(nrow(wdbc), fit))  # plot simulated data

}
\seealso{
\code{\link{kdevine}},
\code{\link{rkdevinecop}},
\code{\link{rkde1d}},
\code{\link[qrng:ghalton]{ghalton}}
}

